clean: MainBodyのソースコードを新しくMainBodyScreenファイルを作成して移動
