+++ app/src/main/java/com/example/todo_app_jpc/ui/home/TodoApp.kt
@@ -53,88 +53,6 @@ import com.example.todo_app_jpc.ui.todo.TodoEntryViewModel
 import kotlinx.coroutines.launch
 
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    var text by rememberSaveable { mutableStateOf("") }
-    Column(
-        modifier = modifier,
-
-        ) {
-        Text(
-            text = "Hello $name!",
-            modifier = modifier,
-            color = MaterialTheme.colorScheme.onPrimaryContainer
-
-        )
-        TextField(value = text, onValueChange = { text = it })
-
-    }
-}
-
-//これはtutorialで使われてた画面
-@Composable
-private fun MainScreen(
-    todoEntityList: List<TodoEntity>, onTodoClick: (Int) -> Unit, modifier: Modifier = Modifier
-) {
-    Column(
-        horizontalAlignment = Alignment.CenterHorizontally,
-        modifier = modifier
-    ) {
-        if (todoEntityList.isEmpty()) {
-            Text(
-                text = stringResource(R.string.no_item_description),
-                textAlign = TextAlign.Center,
-                style = MaterialTheme.typography.titleLarge
-            )
-        } else {
-            AllTodoList(
-                todoEntityList = todoEntityList,
-                onTodoClick = { onTodoClick(it.id) },
-                modifier = Modifier.padding(horizontal = dimensionResource(id = R.dimen.padding_small))
-            )
-        }
-    }
-}
-
-//　これもtutorialのやつ
-@Composable
-private fun AllTodoList(
-    todoEntityList: List<TodoEntity>, onTodoClick: (TodoEntity) -> Unit, modifier: Modifier = Modifier
-) {
-    LazyColumn(modifier = modifier){
-        items(items = todoEntityList, key = {it.id}) { item ->
-            TodoItem(todoEntity = item, modifier = Modifier
-                .padding(dimensionResource(id = R.dimen.padding_small))
-                .clickable { onTodoClick(item) })
-        }
-    }
-}
-
-@Composable
-private fun TodoItem(
-    todoEntity: TodoEntity, modifier: Modifier = Modifier
-) {
-    Card(
-        modifier = modifier, elevation = CardDefaults.cardElevation(defaultElevation = 2.dp)
-    ) {
-        Column(
-            modifier = Modifier.padding(dimensionResource(id = R.dimen.padding_large)),
-            verticalArrangement = Arrangement.spacedBy(dimensionResource(id = R.dimen.padding_small))
-        ) {
-//            rowいらない説濃厚
-            Row(
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Text(
-                    text = todoEntity.title,
-                    style = MaterialTheme.typography.titleLarge,
-                )
-
-            }
-        }
-    }
-}
